Expand args.loss_related_use_float64

+++ library/sdxl_train_util.py
@@ -179,7 +179,7 @@ def match_mixed_precision(args, weight_dtype):
         return None
 
 
-def timestep_embedding(timesteps, dim, max_period=10000):
+def timestep_embedding(timesteps, dim, max_period=10000, dtype=torch.float32):
     """
     Create sinusoidal timestep embeddings.
     :param timesteps: a 1-D Tensor of N indices, one per batch element.
@@ -189,29 +189,29 @@ def timestep_embedding(timesteps, dim, max_period=10000):
     :return: an [N x dim] Tensor of positional embeddings.
     """
     half = dim // 2
-    freqs = torch.exp(-math.log(max_period) * torch.arange(start=0, end=half, dtype=torch.float32) / half).to(
+    freqs = torch.exp(-math.log(max_period) * torch.arange(start=0, end=half, dtype=dtype) / half).to(
         device=timesteps.device
     )
-    args = timesteps[:, None].float() * freqs[None]
+    args = timesteps[:, None].to(dtype=dtype) * freqs[None]
     embedding = torch.cat([torch.cos(args), torch.sin(args)], dim=-1)
     if dim % 2:
         embedding = torch.cat([embedding, torch.zeros_like(embedding[:, :1])], dim=-1)
     return embedding
 
 
-def get_timestep_embedding(x, outdim):
+def get_timestep_embedding(x, outdim, dtype=torch.float32):
     assert len(x.shape) == 2
     b, dims = x.shape[0], x.shape[1]
     x = torch.flatten(x)
-    emb = timestep_embedding(x, outdim)
+    emb = timestep_embedding(x, outdim, dtype=dtype)
     emb = torch.reshape(emb, (b, dims * outdim))
     return emb
 
 
-def get_size_embeddings(orig_size, crop_size, target_size, device):
-    emb1 = get_timestep_embedding(orig_size, 256)
-    emb2 = get_timestep_embedding(crop_size, 256)
-    emb3 = get_timestep_embedding(target_size, 256)
+def get_size_embeddings(orig_size, crop_size, target_size, device, dtype=torch.float32):
+    emb1 = get_timestep_embedding(orig_size, 256, dtype=dtype)
+    emb2 = get_timestep_embedding(crop_size, 256, dtype=dtype)
+    emb3 = get_timestep_embedding(target_size, 256, dtype=dtype)
     vector = torch.cat([emb1, emb2, emb3], dim=1).to(device)
     return vector
